data/reporting/requirements.txt
appnope==0.1.2
APScheduler==3.6.3
autopep8==1.5.7
backcall==0.2.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
idna==3.3
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==7.0.6
jupyter-core==4.8.1
kiwisolver==1.3.2
matplotlib==3.4.3
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
numpy==1.22.4
pandas==1.3.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
PiHole-api==2.6
Pillow==8.4.0
prompt-toolkit==3.0.21
ptyprocess==0.7.0
pycodestyle==2.8.0
Pygments==2.10.0
pyparsing==3.0.0
python-dateutil==2.8.2
python-decouple==3.5
python-telegram-bot==13.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
requests==2.26.0
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
toml==0.10.2
tornado==6.1
traitlets==5.1.0
tzdata==2021.4
tzlocal==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
discord_webhook==0.16.3
